requirements.txt
numpy
matplotlib
scipy
pip
customtkinter
thorlabs-elliptec
pyserial
py_pol
